Adds 'display all users' button to editCredits.php
Adds a 'DISPLAY ALL USERS" button to the Credit Manageement page, which lives on editCredits.php. Queries using an empty string, which should always work.

Additionally, fixed a bug where the return search query from process_editCredits.php was not encoding the searchQuery, which resulted in the page not loading properly on empty query searches.